refactor to routing.h
