+++ core/core.cbp
@@ -98,6 +98,7 @@
 		<Unit filename="pt_ss_k.h" />
 		<Unit filename="pt_ss_k_cell_model.h" />
 		<Unit filename="region_model.h" />
+		<Unit filename="routing.h" />
 		<Unit filename="sceua_optimizer.cpp" />
 		<Unit filename="sceua_optimizer.h" />
 		<Unit filename="skaugen.h" />
